models/sampling/confidence_guided_sampler.py
"""
Confidence-Guided Sampling implementation for ProteinFlex.
Based on recent advances in protein structure generation.
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
import numpy as np
from typing import Dict, Tuple, Optional

class ConfidenceGuidedSampler(nn.Module):
    def __init__(
        self,
        feature_dim: int = 768,
        hidden_dim: int = 512,
        num_steps: int = 1000,
        min_beta: float = 1e-4,
        max_beta: float = 0.02
    ):
        """
        Initialize the Confidence-Guided Sampler.

        Args:
            feature_dim: Dimension of protein features
            hidden_dim: Hidden dimension for confidence network
            num_steps: Number of diffusion steps
            min_beta: Minimum noise schedule value
            max_beta: Maximum noise schedule value
        """
        super().__init__()
        self.feature_dim = feature_dim
        self.hidden_dim = hidden_dim

        # Confidence estimation network
        self.confidence_net = nn.Sequential(
            nn.Linear(feature_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, hidden_dim // 2),
            nn.LayerNorm(hidden_dim // 2),
            nn.ReLU(),
            nn.Linear(hidden_dim // 2, 1),
            nn.Sigmoid()
        )

        # Noise prediction network
        self.noise_pred_net = nn.Sequential(
            nn.Linear(feature_dim + hidden_dim, hidden_dim),  # +hidden_dim for time embedding
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, feature_dim)
        )

        # Setup noise schedule
        self.num_steps = num_steps
        self.register_buffer('betas', torch.linspace(min_beta, max_beta, num_steps))
        alphas = 1 - self.betas
        self.register_buffer('alphas_cumprod', torch.cumprod(alphas, dim=0))

    def get_time_embedding(self, t: torch.Tensor) -> torch.Tensor:
        """Generate time embeddings."""
        half_dim = self.hidden_dim // 2
        embeddings = torch.log(torch.tensor(10000.0)) / (half_dim - 1)
        embeddings = torch.exp(torch.arange(half_dim, device=t.device) * -embeddings)
        embeddings = t[:, None] * embeddings[None, :]
        embeddings = torch.cat((embeddings.sin(), embeddings.cos()), dim=-1)
        return embeddings

    def forward(
        self,
        x: torch.Tensor,
        noise: Optional[torch.Tensor] = None
    ) -> Tuple[torch.Tensor, torch.Tensor]:
        """
        Forward pass for training.

        Args:
            x: Input protein features [batch_size, seq_len, feature_dim]
            noise: Optional pre-generated noise

        Returns:
            Tuple of (noisy features, predicted noise)
        """
        batch_size = x.shape[0]

        # Generate noise if not provided
        if noise is None:
            noise = torch.randn_like(x)

        # Sample timestep
        t = torch.randint(0, self.num_steps, (batch_size,), device=x.device)

        # Get noise scaling
        a = self.alphas_cumprod[t]
        a = a.view(-1, 1, 1)

        # Add noise to input
        noisy_x = torch.sqrt(a) * x + torch.sqrt(1 - a) * noise

        # Predict noise
        time_emb = self.get_time_embedding(t)
        time_emb = time_emb.view(batch_size, 1, -1).expand(-1, x.shape[1], -1)
        pred_input = torch.cat([noisy_x, time_emb], dim=-1)
        pred_noise = self.noise_pred_net(pred_input)

        return noisy_x, pred_noise

    def sample(
        self,
        batch_size: int,
        seq_len: int,
        device: torch.device,
        temperature: float = 1.0
    ) -> torch.Tensor:
        """
        Generate protein features using confidence-guided sampling.

        Args:
            batch_size: Number of samples to generate
            seq_len: Sequence length
            device: Device to generate on
            temperature: Sampling temperature

        Returns:
            Generated protein features
        """
        # Start from random noise
        x = torch.randn(batch_size, seq_len, self.feature_dim, device=device)

        # Iterative refinement
        for t in reversed(range(self.num_steps)):
            # Get confidence score
            confidence = self.confidence_net(x)

            # Predict and remove noise
            time_emb = self.get_time_embedding(torch.tensor([t], device=device))
            time_emb = time_emb.expand(batch_size, seq_len, -1)
            pred_input = torch.cat([x, time_emb], dim=-1)
            pred_noise = self.noise_pred_net(pred_input)

            # Update features based on confidence
            alpha = self.alphas_cumprod[t]
            alpha_prev = self.alphas_cumprod[t-1] if t > 0 else torch.tensor(1.0, device=device)
            beta = 1 - alpha / alpha_prev

            # Apply confidence-guided update
            mean = (x - beta * pred_noise) / torch.sqrt(1 - beta)
            var = beta * temperature * (1 - confidence)
            x = mean + torch.sqrt(var) * torch.randn_like(x)

        return x

    def compute_loss(
        self,
        x: torch.Tensor,
        noise: torch.Tensor,
        pred_noise: torch.Tensor
    ) -> torch.Tensor:
        """
        Compute training loss.

        Args:
            x: Input protein features
            noise: Target noise
            pred_noise: Predicted noise

        Returns:
            Combined loss value
        """
        # MSE loss for noise prediction
        noise_loss = F.mse_loss(pred_noise, noise)

        # Confidence loss to encourage accurate confidence estimation
        confidence = self.confidence_net(x)
        confidence_target = torch.exp(-F.mse_loss(pred_noise, noise, reduction='none').mean(-1))
        confidence_loss = F.binary_cross_entropy(confidence, confidence_target)

        return noise_loss + confidence_loss
